requirements.txt
yandexcloud>=0.228.0
paramiko>=2.11.0
PyYAML>=6.0
requests>=2.31.0 